% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit.R
\name{add_commit_push}
\alias{add_commit_push}
\title{Add, commit, and push changes to a Git repository}
\usage{
add_commit_push(commit_message, prepend)
}
\arguments{
\item{commit_message}{A character string containing the commit message for
the changes.}

\item{prepend}{A character string to prepend to the commit message.
Defaults to an empty string.}
}
\description{
Aautomates the process of adding changes to a Git repository,
committing
those changes with a commit message, and pushing the changes to a remote
repository.
}
\examples{
\dontrun{
  add_commit_push()
}
}
