% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit.R
\name{generate_commit_message}
\alias{generate_commit_message}
\title{Generate commit message}
\usage{
generate_commit_message(encoded_git_diff_output)
}
\arguments{
\item{encoded_git_diff_output}{A character vector containing the URL-encoded
git diff output.}
}
\value{
A character vector of length 1 containing the generated commit
    message.
}
\description{
Generates a commit message by making an API call to OpenAI's
GPT-3.5 Turbo model using the
encoded git diff output as input.
}
\examples{
\dontrun{
  # Sends the encoded git diff to GPT and returns a
  # character vector containing the commit message:
  git_diff_output <- generate_git_diff_output()
  edo <- generate_encoded_git_diff_output(git_diff_output)
  enerate_commit_message(edo)
}
}
