% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit.R
\name{generate_encoded_git_diff_output}
\alias{generate_encoded_git_diff_output}
\title{Encode git diff output}
\usage{
generate_encoded_git_diff_output(git_diff_output)
}
\arguments{
\item{git_diff_output}{A character vector containing the git diff output.}
}
\value{
A character vector containing the URL-encoded git diff output.
}
\description{
Takes the git diff output and encodes it for use in the API
call to the 'OpenAI' 'GPT-3.5 Turbo' model.
}
\examples{
\dontrun{
  git_diff_output <- generate_git_diff_output()
  generate_encoded_git_diff_output(git_diff_output)
}
}
