% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit.R
\name{generate_git_diff_output}
\alias{generate_git_diff_output}
\title{Generate git diff output}
\usage{
generate_git_diff_output()
}
\value{
A character vector containing the git diff output.
}
\description{
Returns the git diff output for the current working directory, and lists
any new files found.
It checks the system's OS and executes the appropriate command to generate
the git diff.
}
\examples{
\dontrun{
  # View with cat() for easier reading
  cat(gitGPT:::generate_git_diff_output())
}
}
