% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit.R
\name{suggest_commit_message}
\alias{suggest_commit_message}
\title{Suggest a commit message based on the provided git diff}
\usage{
suggest_commit_message(diff)
}
\arguments{
\item{diff}{An optional character vector containing the git diff. If not
provided, the function will automatically
generate the git diff output for the current working directory.}
}
\value{
A character vector of length 1 with the suggested commit message
    based on the provided git diff.
}
\description{
Suggests a commit message by utilizing the 'OpenAI' 'GPT-3.5
Turbo' model.
It takes a git diff as input and returns a meaningful
commit message.
}
\examples{
\dontrun{
  suggest_commit_message()
}
}
